COMMON: Move huffman.h to common/compression/
